♻️ Extract public/id_gen*.go from id_gen*.go and add aliases
