+++ id_gen.go
@@ -1,22 +1,8 @@
 package sqlslog
 
-// IDGen is a function that generates an ID string.
-type IDGen = func() string
-
 // IDGenerator returns an Option that sets the ID generator.
 // The default is IDGeneratorDefault.
 func IDGenerator(idGen IDGen) Option { return func(o *Options) { o.idGen = idGen } }
-
-const defaultIDLength = 16
-
-// IDGeneratorDefault is the default ID generator.
-var IDGeneratorDefault = NewChaCha8IDGenerator(defaultIDLength).Generate
-
-const (
-	ConnIDKeyDefault = "conn_id"
-	TxIDKeyDefault   = "tx_id"
-	StmtIDKeyDefault = "stmt_id"
-)
 
 // ConnIDKey sets the key for the connection ID.
 // The default is ConnIDKeyDefault.
