+++ public/id_gen_chacha8.go
@@ -1,4 +1,4 @@
-package sqlslog
+package public
 
 import (
 	cryptoRand "crypto/rand"
